rename test file and mark run as complete when test is done
